move fa_job.py to jobs/_fa_job.py
