+++ fedbiomed/node/jobs/__init__.py
@@ -1,0 +1,12 @@
+# This file is originally part of Fed-BioMed
+# SPDX-License-Identifier: Apache-2.0
+
+"""
+Classes that simplify imports from fedbiomed.node.jobs package
+"""
+
+from ._fa_job import FAJob
+
+__all__ = [
+    "FAJob",
+]
